Kinematics::Inverse loop cleanup
Moved the initial call to _MyTrig and the initial setup of the criteria variables inside the while loop and eliminated them at the end of the loop. This eliminates 1 call to each of the required functions per call to Inverse. Essentially this eliminates an unneccessary call to these functions at the conclusion of the final interation of the loop.